refactor(dev): rename serverjs namespace to `server`
