+++ packages/remix-dev/compiler/server/plugins/manifest.ts
