--- packages/remix-dev/compiler/serverjs/index.ts
@@ -1,1 +1,0 @@
-export * as compiler from "./compiler";
